Update control-access-overview.md
You can control the crypto-usage (ex. Encrypt, Decrypt) of a KMS key through an IAM policy attached to a user, even without explicitly adding them as a KMS CMK Key-User, and doing so with the default (non-modified) Key-Policy that is in place. In other words, once the KMS CMK key is created through the GUI console (as well as its default Key-Policy), it is possible to assign a user or a role the IAM permissions to encrypt/decrypt data. The documentation seems to say imply that Key-Policy should allow that access, regardless of any IAM policies. 

To test: create a CMK with all the defaults, and without adding any Admin or any Key-Users. Create a new IAM user, and assign them the following: 
    "kms:Decrypt",
    "kms:DescribeKey",
    "kms:Encrypt",
    "kms:GenerateDataKey*",
    "kms:ReEncrypt*"

That IAM user would be able to encrypt data, through the CLI, even though the default Key-Policy was not modified, and even though they were never added under the Key Users section of the CMK. If this behavior is normal, I believe the documentation should be changed to reflect all of this.